- make targinfo[2] uint8_t rather than int.
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox@21134 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ branches/RATBOX_2_1_MEMTIDY/include/client.h
@@ -254,7 +254,7 @@ struct LocalUser
 
 	/* target change stuff */
 	void *targets[10];		/* targets were aware of */
-	unsigned int targinfo[2];	/* cyclic array, no in use */
+	uint8_t targinfo[2];		/* position in cyclic array, no in use */
 	time_t target_last;		/* last time we cleared a slot */
 };
